% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipenrich_package_doc.R
\docType{package}
\name{chipenrich_package}
\alias{chipenrich_package}
\alias{chipenrich_package-package}
\title{chipenrich: Gene Set Enrichment For ChIP-seq Peak Data}
\description{
The \code{chipenrich} package includes three classes of methods that adjust for potential confounders of gene set enrichment testing (locus length and mappability of the sequence reads). The first, \code{chipenrich}, is designed for use with transcription-factor (TF) based ChIP-seq experiments and other DNA sequencing experiments with narrow genomic regions. The second, \code{polyenrich}, is similarly designed for TF based ChIP-seq, but where the number of peaks present in gene loci may be important. The third, \code{broadenrich}, is designed for use with histone modification based ChIP-seq experiments and other DNA sequencing experiments with broad genomic regions.
}
