% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{filter_genesets}
\alias{filter_genesets}
\title{Function to filter genesets from being too large}
\usage{
filter_genesets(gs_obj, max_geneset_size = 2000)
}
\arguments{
\item{gs_obj}{A valid GeneSet object}

\item{max_geneset_size}{An integer indicating the ceiling for genes in a geneset. Default 2000.}
}
\value{
An altered \code{gs_obj} with changed \code{set.gene} and \code{all.genes} slots reflecting \code{max_geneset_size}.
}
\description{
Function to filter genesets from being too large
}
