% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genome_to_organism}
\alias{genome_to_organism}
\title{Get the correct organism code based on genome}
\usage{
genome_to_organism(genome = supported_genomes())
}
\arguments{
\item{genome}{One of the \code{supported_genomes()}.}
}
\value{
A string for the three letter organism code. Convention is first letter
of the first word in the binomial name, and first two letters of the second word
in the binomial name. 'Homo sapiens' is then 'hsa', for example.
}
\description{
Data from \code{chipenrich.data} uses three letter organism codes for the
\code{GeneSet} objects. This function ensures the correct objects are loaded.
}
