% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genome_to_orgdb}
\alias{genome_to_orgdb}
\title{Get Entrez ID to gene symbol mappings for custom locus definitions}
\usage{
genome_to_orgdb(genome = supported_genomes())
}
\arguments{
\item{genome}{One of the \code{supported_genomes()}.}
}
\value{
A \code{data.frame} with \code{gene_id} and \code{symbol} columns.
}
\description{
If a custom locus definition is one of the \code{supported_genomes()}, then
the gene symbol column of the custom locus definition is populated using the
appropriate orgDb package.
}
