% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_test_method}
\alias{get_test_method}
\title{Get the test function name from the method name}
\usage{
get_test_method(method)
}
\arguments{
\item{method}{A character for the method used. One of the \code{supported_methods}
or one of the \code{HIDDEN_METHODS} in \code{constants.R}.}
}
\value{
A singleton named character vector with value of the test function and
name of the method.
}
\description{
The \code{method} comes from what is used in \code{chipenrich()} or in
\code{polyenrich()}.
}
