% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hybridenrich}
\alias{hybridenrich}
\title{Running Hybrid test, either from scratch or using two results files}
\usage{
hybridenrich(peaks, out_name = "hybridenrich", out_path = getwd(),
  genome = supported_genomes(), genesets = c("GOBP", "GOCC", "GOMF"),
  locusdef = "nearest_tss", methods = c("chipenrich", "polyenrich"),
  weighting = NULL, mappability = NULL, qc_plots = TRUE,
  min_geneset_size = 15, max_geneset_size = 2000, num_peak_threshold = 1,
  randomization = NULL, n_cores = 1)
}
\description{
Hybrid test is designed for people unsure of which test between ChIP-Enrich
and Poly-Enrich to use, so it takes information of both and gives adjusted
P-values. For more about ChIP- and Poly-Enrich, consult their corresponding
documentation.
}
\section{Hybrid p-values}{

Given n tests that test for the same hypothesis, same Type I error rate, and
converted to p-values: \code{p_1, ..., p_n}, the Hybrid p-value is computed as:
\code{n*min(p_1, ..., p_n)}. This hybrid test will have at most the same
Type I error as any individual test, and if any of the tests have 100% power as
sample size goes to infinity, then so will the hybrid test.
}

\section{Function inputs}{

Every input in hybridenrich is the same as in chipenrich and polyenrich. Inputs
unique to chipenrich are: num_peak_threshold; and inputs unique to polyenrich are:
weighting. Currently the test only supports running chipenrich and polyenrich, but
future plans will allow you to run any number of different support tests.
}

