% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{load_peaks}
\alias{load_peaks}
\title{Convert BEDX+Y data.frames and into GRanges}
\usage{
load_peaks(dframe)
}
\arguments{
\item{dframe}{A BEDX+Y style \code{data.frame}.}
}
\value{
A \code{GRanges} that is unstranded, and contains only chrom, start, and end.
}
\description{
Given a \code{data.frame} in BEDX+Y format, keep only the first three
columns: chrom, start, end, and output \code{GRanges}.
}
\details{
Typically, this function will not be used alone, but inside \code{chipenrich()}.
}
\examples{

# Example of BED3 with no header
data(peaks_H3K4me3_GM12878, package='chipenrich.data')
peaks = load_peaks(peaks_H3K4me3_GM12878)

}

