% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chipenrich_fast.R
\name{multi_chipenrich}
\alias{multi_chipenrich}
\title{faster version of 'single_chipenrich' that uses vectorised code'}
\usage{
multi_chipenrich(
  geneset,
  gpw,
  fitspl,
  method,
  model,
  skip_zero = T,
  test_go_ids = NULL,
  sample_size = 10,
  n_cores = 1
)
}
\description{
faster version of 'single_chipenrich' that uses vectorised code'
}
\author{
Brandon Signal
}
\keyword{internal}
