% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post_process_enrichments}
\alias{post_process_enrichments}
\title{Post process the \code{data.frame} of enrichment results}
\usage{
post_process_enrichments(enrich)
}
\arguments{
\item{enrich}{A \code{data.frame} of the enrichment results from \code{broadenrich()},
\code{chipenrich()}, or \code{polyenrich()} created by \code{rbind}ing the
list of enrichment results for each of the \code{genesets}.}
}
\value{
A reformatted \code{data.frame} with columns in a specific order, filtered
of enrichment tests that failed, and ordered first by enrichment 'Status' (if
present) and then 'P.value'.
}
\description{
Post process the \code{data.frame} of enrichment results
}
