% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{postprocess_peak_grs}
\alias{postprocess_peak_grs}
\title{A helper function to post-process peak GRanges}
\usage{
postprocess_peak_grs(gr)
}
\arguments{
\item{gr}{A \code{GRanges} of input peaks.}
}
\value{
A \code{GRanges} that is sorted if the \code{seqinfo} is set, and has named peaks.
}
\description{
Check for overlapping input regions, sort peaks, and force peak names
}
