% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bed}
\alias{read_bed}
\title{Read BEDX+Y files and convert into GRanges}
\usage{
read_bed(file_path)
}
\arguments{
\item{file_path}{A path to a valid BEDX+Y file.}
}
\value{
A \code{GRanges} that is unstranded, and contains only chrom, start, and end.
}
\description{
Given a \code{file_path}, read in a delimited file, assuming it is BEDX+Y,
keep only the first three columns: chrom, start, end, and output \code{GRanges}.
}
\details{
Typically, this function will not be used alone, but inside \code{chipenrich()}.
}
\examples{

# Example of BED3 with no header
file = system.file('extdata', 'test_assign.bed', package = 'chipenrich')
peaks = read_bed(file)

# Example of BED3 with header
file = system.file('extdata', 'test_header.bed', package = 'chipenrich')
peaks = read_bed(file)

# Example of narrowPeak with header
file = system.file('extdata', 'test.narrowPeak', package = 'chipenrich')
peaks = read_bed(file)

}

