% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bedgff}
\alias{read_bedgff}
\title{Read BEDGFF files and convert into GRanges}
\usage{
read_bedgff(file_path)
}
\arguments{
\item{file_path}{A path to a valid BEDGFF file (as from modENCODE).}
}
\value{
A \code{GRanges} that is unstranded, and contains only chrom, start, and end.
}
\description{
Given a \code{file_path}, read in a delimited file, assuming it is BEDGFF (as
is output by modENCODE for D. Melanogaster TF ChIP-seq experiments), keep
only chrom, start, and end columns, and output \code{GRanges}.
}
\details{
Typically, this function will not be used alone, but inside \code{chipenrich()}.
}
\examples{

# Example of GFF3
file = system.file('extdata', 'test.gff3', package = 'chipenrich')
peaks = read_bedgff(file)

# Example of gzipped GFF3
file = system.file('extdata', 'test.gff3.gz', package = 'chipenrich')
peaks = read_bedgff(file)

}

