% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_geneset}
\alias{read_geneset}
\title{Function to read custom gene sets from file}
\usage{
read_geneset(file_path)
}
\arguments{
\item{file_path}{A file path for the custom gene set.}
}
\value{
A \code{GeneSet} class object.
}
\description{
This function reads a two-columned tab-delimited text file (with header). Column
names are ignored, but the first column should be geneset names or IDs and the
second column should be Entrez Gene IDs.
}
