% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mappa}
\alias{read_mappa}
\title{Function to read custom mappability files}
\usage{
read_mappa(file_path)
}
\arguments{
\item{file_path}{A file path for the custom mappability.}
}
\value{
A \code{data.frame} containing \code{gene_id} and \code{mappa} columns.
}
\description{
This function reads a two-columned tab-delimited text file (with header). Expected
column names are 'mappa' and 'gene_id'. Each line is for a unique 'gene_id'
and contains the mappability (between 0 and 1) for that gene.
}
