% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_peaks}
\alias{recode_peaks}
\title{Recode a vector of number of peaks to binary based on threshold}
\usage{
recode_peaks(num_peaks, threshold = 1)
}
\arguments{
\item{num_peaks}{An \code{integer} vector representing numbers of peaks per gene.}

\item{threshold}{An \code{integer} specifying the minimum number of peaks
required to code as 1.}
}
\value{
An binary vector where an entry is 1 if the corresponding entry of
\code{num_peaks} is >= \code{threshold} and is otherwise 0.
}
\description{
Recode a vector of number of peaks to binary based on threshold
}
