% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_ncores_for_windows}
\alias{reset_ncores_for_windows}
\title{Reset n_cores for Windows}
\usage{
reset_ncores_for_windows(n_cores)
}
\arguments{
\item{n_cores}{An \code{integer} passed to \code{broadenrich()},
\code{chipenrich()}, or \code{polyenrich()} indicating the number of cores
to use for enrichment testing.}
}
\value{
Either the original \code{n_cores} if the OS is not Windows, or 1 if
the OS is Windows.
}
\description{
We use \code{parallel::mclapply} for multicore geneset enrichment testing, but
this function supports more than one core if the OS is not Windows. If the OS
is windows, the number of cores (\code{mc.cores}) must be set to 1.
}
