% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_locusdef}
\alias{setup_locusdef}
\title{Function to setup locus definitions}
\usage{
setup_locusdef(ldef_code, genome, randomization = NULL)
}
\arguments{
\item{ldef_code}{One of 'nearest_tss', 'nearest_gene', 'exon', 'intron', '1kb',
'1kb_outside', '1kb_outside_upstream', '5kb', '5kb_outside', '5kb_outside_upstream',
'10kb', '10kb_outside', '10kb_outside_upstream'. Alternately, a file path for
a custom locus definition. NOTE: Must be for a \code{supported_genome()}, and
must have columns 'chr', 'start', 'end', and 'gene_id', or 'geneid'.}

\item{genome}{One of the \code{supported_genomes()}.}

\item{randomization}{One of \code{NULL}, 'complete', 'bylength', or 'bylocation'.
See the Randomizations section in \code{?chipenrich}. Default NULL.}
}
\value{
A list with components \code{ldef} and \code{tss}.
}
\description{
Function to setup locus definitions
}
