% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chipenrich_fast.R
\name{single_chipenrich}
\alias{single_chipenrich}
\title{edited version of 'single_chipenrich' that can run quietly'}
\usage{
single_chipenrich(go_id, geneset, gpw, fitspl, method, model, verbose = T)
}
\description{
edited version of 'single_chipenrich' that can run quietly'
}
\author{
Brandon Signal
}
\keyword{internal}
