% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\name{supported_genesets}
\alias{supported_genesets}
\title{Display supported genesets for gene set enrichment.}
\usage{
supported_genesets()
}
\value{
A vector indicating supported gene sets for enrichment.
}
\description{
Display supported genesets for gene set enrichment.
}
\examples{

supported_genesets()

}

