\name{supported_genomes}
\alias{supported_genomes}

\title{
List supported genomes
}

\description{
Obtain list of supported genomes
}

\usage{
supported_genomes()
}

\value{
	A character vector listing the currently supported genomes by the \code{\link{chipenrich}} function. 
}

\author{
Ryan Welch \email{welchr@umich.edu}
}

\seealso{
	\code{\link{chipenrich}}
}

\examples{
supported_genomes()
}
