\name{supported_methods}
\alias{supported_methods}

\title{
List supported methods (enrichment tests) 
}

\description{
Obtain list of supported methods (enrichment tests) 
}

\usage{
supported_methods()
}

\value{
	A character vector listing the currently supported methods by the \code{\link{chipenrich}} function. 
}

\author{
Ryan Welch \email{welchr@umich.edu}
}

\seealso{
	\code{\link{chipenrich}}
}

\examples{
supported_methods()
}
