% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\name{supported_read_lengths}
\alias{supported_read_lengths}
\title{Display supported read lengths for mappability}
\usage{
supported_read_lengths()
}
\value{
A vector indicating supported mappability read lengths.
}
\description{
Display supported read lengths for mappability
}
\examples{

supported_read_lengths()

}

