% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chipenrich_fast.R
\name{test_chipenrich_fast}
\alias{test_chipenrich_fast}
\title{faster version of 'test_chipenrich'}
\usage{
test_chipenrich_fast(
  geneset,
  gpw,
  n_cores = 1,
  gpw.bg = NULL,
  bg_peaks = NULL,
  pre_filter = TRUE,
  genome,
  peak
)
}
\description{
faster version of 'test_chipenrich'
}
\author{
Brandon Signal
}
\keyword{internal}
