% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_to_tss.R
\name{plot_dist_to_tss}
\alias{plot_dist_to_tss}
\title{Plot histogram of distance from peak to nearest TSS}
\usage{
plot_dist_to_tss(peaks, genome = "hg19")
}
\arguments{
\item{peaks}{A \code{data.frame}, or tab-delimited text file (BED, narrowPeak,
broadPeak, etc) with the first three columns being chrom, start, and end. The
data frame should have at least 3 columns: chrom, start, and end. Chrom
should follow UCSC convention, e.g. "chrX".}

\item{genome}{A string indicating the genome upon which the peaks file is
based. Supported genomes are listed by the \code{\link{supported_genomes}} function.}
}
\value{
A trellis plot object.
}
\description{
Create a histogram of the distance from each peak to the nearest transcription
start site (TSS) of any gene.
}
\examples{

# Create histogram of distance from peaks to nearest TSS.
data(peaks_E2F4, package = 'chipenrich.data')
plot_dist_to_tss(peaks_E2F4)

}

