% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_to_tss.R
\name{plot_dist_to_tss}
\alias{plot_dist_to_tss}
\title{Plot histogram of distance from peak to nearest TSS}
\usage{
plot_dist_to_tss(peaks, genome = supported_genomes())
}
\arguments{
\item{peaks}{A \code{data.frame}, or tab-delimited text file (BED, narrowPeak,
broadPeak, etc) with the first three columns being chrom, start, and end. The
data frame should have at least 3 columns: chrom, start, and end. Chrom
should follow UCSC convention, e.g. "chrX".}

\item{genome}{One of the \code{supported_genomes()}.}
}
\value{
A trellis plot object.
}
\description{
Create a histogram of the distance from each peak to the nearest transcription
start site (TSS) of any gene.
}
\examples{

# Create histogram of distance from peaks to nearest TSS.
data(peaks_E2F4, package = 'chipenrich.data')
peaks_E2F4 = subset(peaks_E2F4, peaks_E2F4$chrom == 'chr1')
plot_dist_to_tss(peaks_E2F4, genome = 'hg19')

}
