\name{supported_genesets}
\alias{supported_genesets}

\title{
List supported genesets
}

\description{
Obtain list of supported genesets
}

\usage{
supported_genesets()
}

\value{
	A character vector listing the currently supported genesets by the \code{\link{chipenrich}} function. 
}

\author{
Ryan Welch \email{welchr@umich.edu}
}

\seealso{
	\code{\link{chipenrich}}
}

\examples{
supported_genesets()
}
