% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\name{supported_genesets}
\alias{supported_genesets}
\title{Display supported genesets for gene set enrichment.}
\usage{
supported_genesets(organism = NULL)
}
\value{
A \code{data.frame} with columns \code{geneset, organism}.
}
\description{
Display supported genesets for gene set enrichment.
}
\examples{

supported_genesets()

}
