% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\name{supported_genomes}
\alias{supported_genomes}
\title{Display supported genomes.}
\usage{
supported_genomes()
}
\value{
A vector indicating supported genomes.
}
\description{
Display supported genomes.
}
\examples{

supported_genomes()

}
