% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\name{supported_locusdefs}
\alias{supported_locusdefs}
\title{Display supported locus definitions}
\usage{
supported_locusdefs()
}
\value{
A vector indicating the supported locus definitions.
}
\description{
Display supported locus definitions
}
\examples{

supported_locusdefs()

}

