\name{supported_locusdefs}
\alias{supported_locusdefs}

\title{
List supported locus definitions
}

\description{
Obtain list of supported locus definitions for assigning peaks to genes
}

\usage{
supported_locusdefs()
}

\value{
	A character vector listing the currently supported locus definitions by the \code{\link{chipenrich}} function. 
}

\author{
Ryan Welch \email{welchr@umich.edu}
}

\seealso{
	\code{\link{chipenrich}}
}

\examples{
supported_locusdefs()
}
