% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\name{supported_methods}
\alias{supported_methods}
\title{Display supported gene set enrichment methods.}
\usage{
supported_methods()
}
\value{
A vector indicating supported methods for gene set enrichment.
}
\description{
Display supported gene set enrichment methods.
}
\examples{

supported_methods()

}

