\name{supported_read_lengths}
\alias{supported_read_lengths}

\title{
List supported read lengths 
}

\description{
Obtain list of supported read lengths (used when adjusting for mappability)
}

\usage{
supported_read_lengths()
}

\value{
	A character vector listing the currently supported mappability read lengths by the \code{\link{chipenrich}} function. 
}

\author{
Ryan Welch \email{welchr@umich.edu}
}

\seealso{
	\code{\link{chipenrich}}
}

\examples{
supported_read_lengths()
}
